org: moved structs related to mp data and map searches to MapDataInputModels
